--- Dockerfile3
@@ -1,4 +1,0 @@
-# syntax=docker/dockerfile:1
-FROM sailenicolas/gophp:8.1.3
-CMD echo $PHP_VERSION_ID
-ENTRYPOINT ["/bin/sh", "-c"]
